Move 0x swap selectors to separate file
