+++ background/redux-slices/0x-swap.ts
@@ -1,14 +1,9 @@
-import {
-  AnyAction,
-  createSelector,
-  createSlice,
-  ThunkDispatch,
-} from "@reduxjs/toolkit"
+import { AnyAction, createSlice, ThunkDispatch } from "@reduxjs/toolkit"
 import { fetchJson } from "@ethersproject/web"
 import { BigNumber, ethers, utils } from "ethers"
 
 import { createBackgroundAsyncThunk } from "./utils"
-import { isSmartContractFungibleAsset, SwappableAsset } from "../assets"
+import { SwappableAsset } from "../assets"
 import logger from "../lib/logger"
 import {
   isValidSwapPriceResponse,
@@ -24,19 +19,15 @@ import {
   OPTIMISM,
   OPTIMISTIC_ETH,
 } from "../constants"
-import { EVMNetwork, sameNetwork } from "../networks"
+import { EVMNetwork } from "../networks"
 import { setSnackbarMessage } from "./ui"
-import {
-  enrichAssetAmountWithDecimalValues,
-  isBuiltInNetworkBaseAsset,
-} from "./utils/asset-utils"
+import { enrichAssetAmountWithDecimalValues } from "./utils/asset-utils"
 import { AssetsState } from "./assets"
 import {
   checkCurrencyAmount,
   PriceDetails,
   SwapQuoteRequest,
 } from "./utils/0x-swap-utils"
-import { selectCurrentNetwork } from "./selectors/uiSelectors"
 
 // This is how 0x represents native token addresses
 const ZEROEX_NATIVE_TOKEN_CONTRACT_ADDRESS =
@@ -249,7 +240,7 @@ export const fetchSwapQuote = createBackgroundAsyncThunk(
       takerAddress: tradeAddress,
     })
 
-    const apiData: ZrxQuote = await fetchJson({
+    const apiData: unknown = await fetchJson({
       url: requestUrl.toString(),
       headers: gatedHeaders,
     })
@@ -499,37 +490,6 @@ export const executeSwap = createBackgroundAsyncThunk(
         timestamp: Date.now(),
         blockTimestamp: undefined,
       },
-    })
-  }
-)
-
-export const selectLatestQuoteRequest = createSelector(
-  (state: { swap: SwapState }) => state.swap.latestQuoteRequest,
-  (latestQuoteRequest) => latestQuoteRequest
-)
-
-export const selectInProgressApprovalContract = createSelector(
-  (state: { swap: SwapState }) => state.swap.inProgressApprovalContract,
-  (approvalInProgress) => approvalInProgress
-)
-
-export const selectSwapBuyAssets = createSelector(
-  (state: { assets: AssetsState }) => state.assets,
-  selectCurrentNetwork,
-  (assets, currentNetwork) => {
-    return assets.filter((asset) => {
-      if (isSmartContractFungibleAsset(asset)) {
-        if (sameNetwork(asset.homeNetwork, currentNetwork)) {
-          return true
-        }
-      }
-      if (
-        // Explicitly add a network's base asset.
-        isBuiltInNetworkBaseAsset(asset, currentNetwork)
-      ) {
-        return true
-      }
-      return false
     })
   }
 )
